third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-active-state-iframe-expected.txt

Tests that tap gesture events set and clear the active state of elements, even when inside an iframe and the document is scrolled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

See https://bugs.webkit.org/show_bug.cgi?id=96060 for details

Verify active isn't initially set
PASS getHoverActiveState(box) is "default"
Verify showPress, tap sets and clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
PASS successfullyParsed is true

TEST COMPLETE
